% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethics.R
\name{anonymise_social_data}
\alias{anonymise_social_data}
\title{Remove common user identifiers from social data}
\usage{
anonymise_social_data(
  x,
  id_cols = c("user", "username", "author", "user_id"),
  method = c("remove", "hash")
)
}
\arguments{
\item{x}{Data frame.}

\item{id_cols}{Columns to remove or hash.}

\item{method}{remove or hash.}
}
\value{
Data frame.
}
\description{
Remove common user identifiers from social data
}
