% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as_culturomic_tbl}
\alias{as_culturomic_tbl}
\title{Coerce data to a culturomic trace table}
\usage{
as_culturomic_tbl(x, dictionary = NULL)
}
\arguments{
\item{x}{Data frame containing digital trace observations.}

\item{dictionary}{Optional dictionary attached as an attribute.}
}
\value{
A culturomic_tbl object.
}
\description{
Coerce data to a culturomic trace table
}
