% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{check_live_dependencies}
\alias{check_live_dependencies}
\title{Check dependencies for live collection}
\usage{
check_live_dependencies(
  platforms = c("wikipedia", "gdelt", "google", "reddit", "youtube")
)
}
\arguments{
\item{platforms}{Platforms to check.}
}
\value{
Data frame with dependency status.
}
\description{
Check dependencies for live collection
}
