% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{culturomic_project}
\alias{culturomic_project}
\title{Create a conservation culturomics project}
\usage{
culturomic_project(name, dictionary = NULL, traces = NULL, events = NULL)
}
\arguments{
\item{name}{Project name.}

\item{dictionary}{Optional concept dictionary.}

\item{traces}{Optional culturomic trace table.}

\item{events}{Optional event calendar.}
}
\value{
A culturomic_project object.
}
\description{
Create a conservation culturomics project
}
