% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{expand_common_names}
\alias{expand_common_names}
\title{Expand dictionary with synonyms and hashtags}
\usage{
expand_common_names(
  dictionary,
  synonyms = NULL,
  hashtags = TRUE,
  lowercase = FALSE
)
}
\arguments{
\item{dictionary}{A culturomic_dictionary.}

\item{synonyms}{Optional synonym data frame.}

\item{hashtags}{Add hashtag queries for common names and synonyms.}

\item{lowercase}{Convert all queries to lower case.}
}
\value{
Expanded dictionary.
}
\description{
Expand dictionary with synonyms and hashtags
}
