% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{get_social_attention}
\alias{get_social_attention}
\title{Alias for social attention collection}
\usage{
get_social_attention(
  dictionary,
  from,
  to,
  platform = c("reddit", "youtube"),
  ...
)
}
\arguments{
\item{dictionary}{Concept dictionary.}

\item{from}{Start date.}

\item{to}{End date.}

\item{platform}{Social platform to collect.}

\item{...}{Passed to the platform collector.}
}
\value{
culturomic_tbl.
}
\description{
Alias for social attention collection
}
