% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethics.R
\name{platform_terms_report}
\alias{platform_terms_report}
\title{Create a platform terms report scaffold}
\usage{
platform_terms_report(platforms)
}
\arguments{
\item{platforms}{Platforms used.}
}
\value{
Data frame.
}
\description{
Create a platform terms report scaffold
}
