% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attention_gap_rank}
\alias{plot_attention_gap_rank}
\title{Plot ranked attention gap}
\usage{
plot_attention_gap_rank(gap, label_table = NULL, label_col = "common_name")
}
\arguments{
\item{gap}{Output from \code{\link[=attention_gap]{attention_gap()}}.}

\item{label_table}{Optional concept label table.}

\item{label_col}{Label column.}
}
\value{
A ggplot object.
}
\description{
Plot ranked attention gap
}
