% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attention_heatmap}
\alias{plot_attention_heatmap}
\title{Plot attention heatmap}
\usage{
plot_attention_heatmap(
  attention,
  label_table = NULL,
  label_col = "common_name",
  value_col = "attention_index"
)
}
\arguments{
\item{attention}{Attention index table.}

\item{label_table}{Optional label table.}

\item{label_col}{Label column.}

\item{value_col}{Attention column.}
}
\value{
A ggplot object.
}
\description{
Plot attention heatmap
}
