% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attention_timeseries}
\alias{plot_attention_timeseries}
\title{Plot attention time series}
\usage{
plot_attention_timeseries(
  attention,
  label_table = NULL,
  label_col = "common_name",
  top_n = NULL,
  facet = TRUE,
  value_col = "attention_index",
  time_unit = c("day", "week", "month"),
  smooth = FALSE
)
}
\arguments{
\item{attention}{Attention index table.}

\item{label_table}{Optional table with concept labels.}

\item{label_col}{Label column.}

\item{top_n}{Optional number of most attended concepts to show.}

\item{facet}{Whether to facet by concept.}

\item{value_col}{Attention column.}

\item{time_unit}{Date aggregation unit.}

\item{smooth}{Whether to add a smooth trend.}
}
\value{
A ggplot object.
}
\description{
Plot attention time series
}
