% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_burst_counts}
\alias{plot_burst_counts}
\title{Plot burst counts by concept}
\usage{
plot_burst_counts(bursts, label_table = NULL, label_col = "common_name")
}
\arguments{
\item{bursts}{Output from \code{\link[=detect_bursts]{detect_bursts()}}.}

\item{label_table}{Optional concept label table.}

\item{label_col}{Label column.}
}
\value{
A ggplot object.
}
\description{
Plot burst counts by concept
}
