% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bursts}
\alias{plot_bursts}
\title{Plot detected attention bursts}
\usage{
plot_bursts(
  bursts,
  label_table = NULL,
  label_col = "common_name",
  time_unit = c("day", "week", "month")
)
}
\arguments{
\item{bursts}{Output from \code{\link[=detect_bursts]{detect_bursts()}}.}

\item{label_table}{Optional concept labels.}

\item{label_col}{Label column.}

\item{time_unit}{Date aggregation unit.}
}
\value{
A ggplot object.
}
\description{
Plot detected attention bursts
}
