% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_campaign_effect_size}
\alias{plot_campaign_effect_size}
\title{Plot campaign effect sizes}
\usage{
plot_campaign_effect_size(effect, standardised = TRUE)
}
\arguments{
\item{effect}{Output from \code{\link[=campaign_effect]{campaign_effect()}}.}

\item{standardised}{Plot standardised effect when available.}
}
\value{
A ggplot object.
}
\description{
Plot campaign effect sizes
}
