% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_campaign_placebo_distribution}
\alias{plot_campaign_placebo_distribution}
\title{Plot campaign placebo distribution}
\usage{
plot_campaign_placebo_distribution(effect, metric = "average_post_effect")
}
\arguments{
\item{effect}{Output from \code{\link[=campaign_effect]{campaign_effect()}}.}

\item{metric}{Placebo metric to plot.}
}
\value{
A ggplot object.
}
\description{
Plot campaign placebo distribution
}
