% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counterfactual_attention}
\alias{plot_counterfactual_attention}
\title{Plot campaign counterfactual attention}
\usage{
plot_counterfactual_attention(effect)
}
\arguments{
\item{effect}{Output from \code{\link[=campaign_effect]{campaign_effect()}}.}
}
\value{
A ggplot object.
}
\description{
Plot campaign counterfactual attention
}
