% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_culturomic_dashboard}
\alias{plot_culturomic_dashboard}
\title{Plot a simple culturomic dashboard}
\usage{
plot_culturomic_dashboard(x)
}
\arguments{
\item{x}{Data frame or package output object.}
}
\value{
A ggplot object when possible.
}
\description{
Plot a simple culturomic dashboard
}
