% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_platform_disagreement}
\alias{plot_platform_disagreement}
\title{Plot platform disagreement}
\usage{
plot_platform_disagreement(
  attention,
  label_table = NULL,
  label_col = "common_name"
)
}
\arguments{
\item{attention}{Attention index table.}

\item{label_table}{Optional label table.}

\item{label_col}{Label column.}
}
\value{
A ggplot object.
}
\description{
Plot platform disagreement
}
