% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_query_ambiguity}
\alias{plot_query_ambiguity}
\title{Plot query ambiguity audit}
\usage{
plot_query_ambiguity(x)
}
\arguments{
\item{x}{Dictionary or output from \code{\link[=audit_queries]{audit_queries()}}.}
}
\value{
A ggplot object.
}
\description{
Plot query ambiguity audit
}
\examples{
dat <- con_culturomics_example_data()
dict <- score_ambiguity(dat$dictionary)
if (requireNamespace("ggplot2", quietly = TRUE)) plot_query_ambiguity(dict)
}
