% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_query_ambiguity_top}
\alias{plot_query_ambiguity_top}
\title{Plot high-ambiguity queries}
\usage{
plot_query_ambiguity_top(x, top_n = 20)
}
\arguments{
\item{x}{Dictionary or output from \code{\link[=audit_queries]{audit_queries()}}.}

\item{top_n}{Number of queries to show.}
}
\value{
A ggplot object.
}
\description{
Plot high-ambiguity queries
}
