% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_query_sensitivity}
\alias{plot_query_sensitivity}
\title{Plot query sensitivity}
\usage{
plot_query_sensitivity(
  sensitivity,
  label_table = NULL,
  label_col = "common_name",
  metric = "query_sensitivity_sd"
)
}
\arguments{
\item{sensitivity}{Output from \code{\link[=query_sensitivity]{query_sensitivity()}}.}

\item{label_table}{Optional table with concept labels.}

\item{label_col}{Label column.}

\item{metric}{Sensitivity column to plot.}
}
\value{
A ggplot object.
}
\description{
Plot query sensitivity
}
