% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_query_type_mix}
\alias{plot_query_type_mix}
\title{Plot query-type composition}
\usage{
plot_query_type_mix(dictionary)
}
\arguments{
\item{dictionary}{Concept dictionary.}
}
\value{
A ggplot object.
}
\description{
Plot query-type composition
}
