% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_seasonality}
\alias{plot_seasonality}
\title{Plot seasonality anomalies}
\usage{
plot_seasonality(seasonality, label_table = NULL, label_col = "common_name")
}
\arguments{
\item{seasonality}{Output from \code{\link[=decompose_seasonality]{decompose_seasonality()}}.}

\item{label_table}{Optional label table.}

\item{label_col}{Label column.}
}
\value{
A ggplot object.
}
\description{
Plot seasonality anomalies
}
