% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sentiment}
\alias{plot_sentiment}
\title{Plot sentiment by concept}
\usage{
plot_sentiment(
  sentiment,
  label_table = NULL,
  label_col = "common_name",
  metric = NULL
)
}
\arguments{
\item{sentiment}{Output from \code{\link[=sentiment_by_concept]{sentiment_by_concept()}}.}

\item{label_table}{Optional concept label table.}

\item{label_col}{Label column.}

\item{metric}{Sentiment metric to plot.}
}
\value{
A ggplot object.
}
\description{
Plot sentiment by concept
}
