% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{query_sensitivity}
\alias{query_sensitivity}
\title{Query sensitivity analysis}
\usage{
query_sensitivity(x, value_col = "scaled_value")
}
\arguments{
\item{x}{culturomic_tbl or standardised trace table.}

\item{value_col}{Value column.}
}
\value{
Data frame summarising query-level instability.
}
\description{
Query sensitivity analysis
}
