% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{validate_queries}
\alias{validate_queries}
\title{Validate query table}
\usage{
validate_queries(dictionary, max_ambiguity = 0.6)
}
\arguments{
\item{dictionary}{A culturomic_dictionary.}

\item{max_ambiguity}{Threshold for warnings.}
}
\value{
Dictionary with validation attributes.
}
\description{
Validate query table
}
