% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{write_culturomic_report}
\alias{write_culturomic_report}
\title{Write culturomics report}
\usage{
write_culturomic_report(x, file, title = "Conservation culturomics report")
}
\arguments{
\item{x}{Result object.}

\item{file}{Output file.}

\item{title}{Report title.}
}
\value{
File path invisibly.
}
\description{
Write culturomics report
}
